--- a/.gitignore
+++ b/.gitignore
@@ -29,3 +29,4 @@ project.lock.json
 .build/
 .testPublish/
 global.json
+BenchmarkDotNet.Artifacts/