--- a/src/Microsoft.Extensions.Logging.Abstractions/LoggerMessage.cs
+++ b/src/Microsoft.Extensions.Logging.Abstractions/LoggerMessage.cs
@@ -245,11 +245,9 @@ namespace Microsoft.Extensions.Logging
             return logValuesFormatter;
         }
 
-        private class LogValues : IReadOnlyList<KeyValuePair<string, object>>
+        private readonly struct LogValues : IReadOnlyList<KeyValuePair<string, object>>
         {
-            public static Func<object, Exception, string> Callback = (state, exception) => ((LogValues)state)._formatter.Format(((LogValues)state).ToArray());
-
-            private static object[] _valueArray = new object[0];
+            public static readonly Func<LogValues, Exception, string> Callback = (state, exception) => state.ToString();
 
             private readonly LogValuesFormatter _formatter;
 
@@ -270,22 +268,14 @@ namespace Microsoft.Extensions.Logging
                 }
             }
 
-            public int Count
-            {
-                get
-                {
-                    return 1;
-                }
-            }
+            public int Count => 1;
 
             public IEnumerator<KeyValuePair<string, object>> GetEnumerator()
             {
                 yield return this[0];
             }
 
-            public object[] ToArray() => _valueArray;
-
-            public override string ToString() => _formatter.Format(ToArray());
+            public override string ToString() => _formatter.Format();
 
             IEnumerator IEnumerable.GetEnumerator()
             {
@@ -293,9 +283,9 @@ namespace Microsoft.Extensions.Logging
             }
         }
 
-        private class LogValues<T0> : IReadOnlyList<KeyValuePair<string, object>>
+        private readonly struct LogValues<T0> : IReadOnlyList<KeyValuePair<string, object>>
         {
-            public static Func<object, Exception, string> Callback = (state, exception) => ((LogValues<T0>)state)._formatter.Format(((LogValues<T0>)state).ToArray());
+            public static readonly Func<LogValues<T0>, Exception, string> Callback = (state, exception) => state.ToString();
 
             private readonly LogValuesFormatter _formatter;
             private readonly T0 _value0;
@@ -310,25 +300,19 @@ namespace Microsoft.Extensions.Logging
             {
                 get
                 {
-                    if (index == 0)
+                    switch (index)
                     {
-                        return new KeyValuePair<string, object>(_formatter.ValueNames[0], _value0);
+                        case 0:
+                            return new KeyValuePair<string, object>(_formatter.ValueNames[0], _value0);
+                        case 1:
+                            return new KeyValuePair<string, object>("{OriginalFormat}", _formatter.OriginalFormat);
+                        default:
+                            throw new IndexOutOfRangeException(nameof(index));
                     }
-                    else if (index == 1)
-                    {
-                        return new KeyValuePair<string, object>("{OriginalFormat}", _formatter.OriginalFormat);
-                    }
-                    throw new IndexOutOfRangeException(nameof(index));
                 }
             }
 
-            public int Count
-            {
-                get
-                {
-                    return 2;
-                }
-            }
+            public int Count => 2;
 
             public IEnumerator<KeyValuePair<string, object>> GetEnumerator()
             {
@@ -338,9 +322,8 @@ namespace Microsoft.Extensions.Logging
                 }
             }
 
-            public object[] ToArray() => new object[] { _value0 };
 
-            public override string ToString() => _formatter.Format(ToArray());
+            public override string ToString() => _formatter.Format(_value0);
 
             IEnumerator IEnumerable.GetEnumerator()
             {
@@ -348,9 +331,9 @@ namespace Microsoft.Extensions.Logging
             }
         }
 
-        private class LogValues<T0, T1> : IReadOnlyList<KeyValuePair<string, object>>
+        private readonly struct LogValues<T0, T1> : IReadOnlyList<KeyValuePair<string, object>>
         {
-            public static Func<object, Exception, string> Callback = (state, exception) => ((LogValues<T0, T1>)state)._formatter.Format(((LogValues<T0, T1>)state).ToArray());
+            public static readonly Func<LogValues<T0, T1>, Exception, string> Callback = (state, exception) => state.ToString();
 
             private readonly LogValuesFormatter _formatter;
             private readonly T0 _value0;
@@ -381,13 +364,7 @@ namespace Microsoft.Extensions.Logging
                 }
             }
 
-            public int Count
-            {
-                get
-                {
-                    return 3;
-                }
-            }
+            public int Count => 3;
 
             public IEnumerator<KeyValuePair<string, object>> GetEnumerator()
             {
@@ -397,9 +374,7 @@ namespace Microsoft.Extensions.Logging
                 }
             }
 
-            public object[] ToArray() => new object[] { _value0, _value1 };
-
-            public override string ToString() => _formatter.Format(ToArray());
+            public override string ToString() => _formatter.Format(_value0, _value1);
 
             IEnumerator IEnumerable.GetEnumerator()
             {
@@ -407,22 +382,16 @@ namespace Microsoft.Extensions.Logging
             }
         }
 
-        private class LogValues<T0, T1, T2> : IReadOnlyList<KeyValuePair<string, object>>
+        private readonly struct LogValues<T0, T1, T2> : IReadOnlyList<KeyValuePair<string, object>>
         {
-            public static Func<object, Exception, string> Callback = (state, exception) => ((LogValues<T0, T1, T2>)state)._formatter.Format(((LogValues<T0, T1, T2>)state).ToArray());
+            public static readonly Func<LogValues<T0, T1, T2>, Exception, string> Callback = (state, exception) => state.ToString();
 
             private readonly LogValuesFormatter _formatter;
-            public T0 _value0;
-            public T1 _value1;
-            public T2 _value2;
+            private readonly T0 _value0;
+            private readonly T1 _value1;
+            private readonly T2 _value2;
 
-            public int Count
-            {
-                get
-                {
-                    return 4;
-                }
-            }
+            public int Count => 4;
 
             public KeyValuePair<string, object> this[int index]
             {
@@ -452,9 +421,7 @@ namespace Microsoft.Extensions.Logging
                 _value2 = value2;
             }
 
-            public object[] ToArray() => new object[] { _value0, _value1, _value2 };
-
-            public override string ToString() => _formatter.Format(ToArray());
+            public override string ToString() => _formatter.Format(_value0, _value1, _value2);
 
             public IEnumerator<KeyValuePair<string, object>> GetEnumerator()
             {
@@ -470,23 +437,17 @@ namespace Microsoft.Extensions.Logging
             }
         }
 
-        private class LogValues<T0, T1, T2, T3> : IReadOnlyList<KeyValuePair<string, object>>
+        private readonly struct LogValues<T0, T1, T2, T3> : IReadOnlyList<KeyValuePair<string, object>>
         {
-            public static Func<object, Exception, string> Callback = (state, exception) => ((LogValues<T0, T1, T2, T3>)state)._formatter.Format(((LogValues<T0, T1, T2, T3>)state).ToArray());
+            public static readonly Func<LogValues<T0, T1, T2, T3>, Exception, string> Callback = (state, exception) => state._formatter.Format(state.ToArray());
 
             private readonly LogValuesFormatter _formatter;
-            public T0 _value0;
-            public T1 _value1;
-            public T2 _value2;
-            public T3 _value3;
+            private readonly T0 _value0;
+            private readonly T1 _value1;
+            private readonly T2 _value2;
+            private readonly T3 _value3;
 
-            public int Count
-            {
-                get
-                {
-                    return 5;
-                }
-            }
+            public int Count => 5;
 
             public KeyValuePair<string, object> this[int index]
             {
@@ -519,7 +480,7 @@ namespace Microsoft.Extensions.Logging
                 _value3 = value3;
             }
 
-            public object[] ToArray() => new object[] { _value0, _value1, _value2, _value3 };
+            private object[] ToArray() => new object[] { _value0, _value1, _value2, _value3 };
 
             public override string ToString() => _formatter.Format(ToArray());
 
@@ -537,24 +498,18 @@ namespace Microsoft.Extensions.Logging
             }
         }
 
-        private class LogValues<T0, T1, T2, T3, T4> : IReadOnlyList<KeyValuePair<string, object>>
+        private readonly struct LogValues<T0, T1, T2, T3, T4> : IReadOnlyList<KeyValuePair<string, object>>
         {
-            public static Func<object, Exception, string> Callback = (state, exception) => ((LogValues<T0, T1, T2, T3, T4>)state)._formatter.Format(((LogValues<T0, T1, T2, T3, T4>)state).ToArray());
+            public static readonly Func<LogValues<T0, T1, T2, T3, T4>, Exception, string> Callback = (state, exception) => state._formatter.Format(state.ToArray());
 
             private readonly LogValuesFormatter _formatter;
-            public T0 _value0;
-            public T1 _value1;
-            public T2 _value2;
-            public T3 _value3;
-            public T4 _value4;
+            private readonly T0 _value0;
+            private readonly T1 _value1;
+            private readonly T2 _value2;
+            private readonly T3 _value3;
+            private readonly T4 _value4;
 
-            public int Count
-            {
-                get
-                {
-                    return 6;
-                }
-            }
+            public int Count => 6;
 
             public KeyValuePair<string, object> this[int index]
             {
@@ -590,7 +545,7 @@ namespace Microsoft.Extensions.Logging
                 _value4 = value4;
             }
 
-            public object[] ToArray() => new object[] { _value0, _value1, _value2, _value3, _value4 };
+            private object[] ToArray() => new object[] { _value0, _value1, _value2, _value3, _value4 };
 
             public override string ToString() => _formatter.Format(ToArray());
 
@@ -608,25 +563,19 @@ namespace Microsoft.Extensions.Logging
             }
         }
 
-        private class LogValues<T0, T1, T2, T3, T4, T5> : IReadOnlyList<KeyValuePair<string, object>>
+        private readonly struct LogValues<T0, T1, T2, T3, T4, T5> : IReadOnlyList<KeyValuePair<string, object>>
         {
-            public static Func<object, Exception, string> Callback = (state, exception) => ((LogValues<T0, T1, T2, T3, T4, T5>)state)._formatter.Format(((LogValues<T0, T1, T2, T3, T4, T5>)state).ToArray());
+            public static readonly Func<LogValues<T0, T1, T2, T3, T4, T5>, Exception, string> Callback = (state, exception) => state._formatter.Format(state.ToArray());
 
             private readonly LogValuesFormatter _formatter;
-            public T0 _value0;
-            public T1 _value1;
-            public T2 _value2;
-            public T3 _value3;
-            public T4 _value4;
-            public T5 _value5;
+            private readonly T0 _value0;
+            private readonly T1 _value1;
+            private readonly T2 _value2;
+            private readonly T3 _value3;
+            private readonly T4 _value4;
+            private readonly T5 _value5;
 
-            public int Count
-            {
-                get
-                {
-                    return 7;
-                }
-            }
+            public int Count => 7;
 
             public KeyValuePair<string, object> this[int index]
             {
@@ -665,86 +614,7 @@ namespace Microsoft.Extensions.Logging
                 _value5 = value5;
             }
 
-            public object[] ToArray() => new object[] { _value0, _value1, _value2, _value3, _value4, _value5 };
-
-            public override string ToString() => _formatter.Format(ToArray());
-
-            public IEnumerator<KeyValuePair<string, object>> GetEnumerator()
-            {
-                for (int i = 0; i < Count; ++i)
-                {
-                    yield return this[i];
-                }
-            }
-
-            IEnumerator IEnumerable.GetEnumerator()
-            {
-                return GetEnumerator();
-            }
-        }
-
-        private class LogValues<T0, T1, T2, T3, T4, T5, T6> : IReadOnlyList<KeyValuePair<string, object>>
-        {
-            public static Func<object, Exception, string> Callback = (state, exception) => ((LogValues<T0, T1, T2, T3, T4, T5, T6>)state)._formatter.Format(((LogValues<T0, T1, T2, T3, T4, T5, T6>)state).ToArray());
-
-            private readonly LogValuesFormatter _formatter;
-            public T0 _value0;
-            public T1 _value1;
-            public T2 _value2;
-            public T3 _value3;
-            public T4 _value4;
-            public T5 _value5;
-            public T6 _value6;
-
-            public int Count
-            {
-                get
-                {
-                    return 8;
-                }
-            }
-
-            public KeyValuePair<string, object> this[int index]
-            {
-                get
-                {
-                    switch (index)
-                    {
-                        case 0:
-                            return new KeyValuePair<string, object>(_formatter.ValueNames[0], _value0);
-                        case 1:
-                            return new KeyValuePair<string, object>(_formatter.ValueNames[1], _value1);
-                        case 2:
-                            return new KeyValuePair<string, object>(_formatter.ValueNames[2], _value2);
-                        case 3:
-                            return new KeyValuePair<string, object>(_formatter.ValueNames[3], _value3);
-                        case 4:
-                            return new KeyValuePair<string, object>(_formatter.ValueNames[4], _value4);
-                        case 5:
-                            return new KeyValuePair<string, object>(_formatter.ValueNames[5], _value5);
-                        case 6:
-                            return new KeyValuePair<string, object>(_formatter.ValueNames[6], _value6);
-                        case 7:
-                            return new KeyValuePair<string, object>("{OriginalFormat}", _formatter.OriginalFormat);
-                        default:
-                            throw new IndexOutOfRangeException(nameof(index));
-                    }
-                }
-            }
-
-            public LogValues(LogValuesFormatter formatter, T0 value0, T1 value1, T2 value2, T3 value3, T4 value4, T5 value5, T6 value6)
-            {
-                _formatter = formatter;
-                _value0 = value0;
-                _value1 = value1;
-                _value2 = value2;
-                _value3 = value3;
-                _value4 = value4;
-                _value5 = value5;
-                _value6 = value6;
-            }
-
-            public object[] ToArray() => new object[] { _value0, _value1, _value2, _value3, _value4, _value5, _value6 };
+            private object[] ToArray() => new object[] { _value0, _value1, _value2, _value3, _value4, _value5 };
 
             public override string ToString() => _formatter.Format(ToArray());
 
@@ -763,4 +633,3 @@ namespace Microsoft.Extensions.Logging
         }
     }
 }
-