1. Make a lot of things struct to avoid boxing
2. Add specialized overloads for 1-3 parameter messages that match `string.Format` and avoid allocation
3. Fix some special casing for no argument logs

### Before
```

                                     Method |      Mean |     Error |    StdDev |         Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
------------------------------------------- |----------:|----------:|----------:|-------------:|-------:|---------:|-------:|----------:|
                NoArguments_FilteredByLevel |  54.30 ns | 0.4572 ns | 0.4053 ns | 18,415,469.9 |   2.72 |     0.02 | 0.0001 |      40 B |
  NoArguments_DefineMessage_FilteredByLevel |  19.94 ns | 0.1230 ns | 0.1027 ns | 50,138,751.2 |   1.00 |     0.00 |      - |       0 B |
                                NoArguments |  74.18 ns | 1.2185 ns | 1.1398 ns | 13,480,512.4 |   3.72 |     0.06 | 0.0001 |      40 B |
                  NoArguments_DefineMessage |  78.33 ns | 1.0545 ns | 0.9348 ns | 12,766,084.6 |   3.93 |     0.05 | 0.0001 |      24 B |
                               TwoArguments | 146.06 ns | 1.7506 ns | 1.6375 ns |  6,846,641.8 |   7.32 |     0.09 | 0.0005 |     104 B |
               TwoArguments_FilteredByLevel | 128.65 ns | 2.5191 ns | 2.8000 ns |  7,772,791.9 |   6.45 |     0.14 | 0.0005 |     104 B |
                 TwoArguments_DefineMessage |  89.52 ns | 0.9914 ns | 0.9274 ns | 11,170,548.1 |   4.49 |     0.05 | 0.0001 |      40 B |
 TwoArguments_DefineMessage_FilteredByLevel |  20.63 ns | 0.2005 ns | 0.1875 ns | 48,463,314.9 |   1.03 |     0.01 |      - |       0 B |


 
       Method |     Mean |    Error |   StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
------------- |---------:|---------:|---------:|------------:|-------:|---------:|-------:|----------:|
 TwoArguments | 355.1 ns | 6.855 ns | 8.161 ns | 2,815,805.2 |   1.65 |     0.04 | 0.0005 |     168 B |
  NoArguments | 215.3 ns | 3.141 ns | 2.785 ns | 4,645,103.7 |   1.00 |     0.00 |      - |      64 B |

```

### After
```
                                     Method |      Mean |     Error |    StdDev |         Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
------------------------------------------- |----------:|----------:|----------:|-------------:|-------:|---------:|-------:|----------:|
                NoArguments_FilteredByLevel |  49.29 ns | 0.5002 ns | 0.4679 ns | 20,286,488.9 |   2.60 |     0.04 |      - |       0 B |
  NoArguments_DefineMessage_FilteredByLevel |  18.96 ns | 0.2912 ns | 0.2724 ns | 52,732,789.4 |   1.00 |     0.00 |      - |       0 B |
                                NoArguments |  64.03 ns | 0.2904 ns | 0.2717 ns | 15,616,744.5 |   3.38 |     0.05 |      - |       0 B |
                  NoArguments_DefineMessage |  53.39 ns | 0.3256 ns | 0.3046 ns | 18,731,554.7 |   2.82 |     0.04 |      - |       0 B |
                               TwoArguments | 134.16 ns | 0.6308 ns | 0.5268 ns |  7,453,628.1 |   7.08 |     0.10 | 0.0002 |      64 B |
               TwoArguments_FilteredByLevel | 124.44 ns | 1.5384 ns | 1.3637 ns |  8,035,928.9 |   6.56 |     0.11 | 0.0002 |      64 B |
                 TwoArguments_DefineMessage |  96.02 ns | 0.6235 ns | 0.5206 ns | 10,414,278.9 |   5.06 |     0.07 |      - |       0 B |
 TwoArguments_DefineMessage_FilteredByLevel |  19.26 ns | 0.1237 ns | 0.1097 ns | 51,914,239.5 |   1.02 |     0.02 |      - |       0 B |

       Method |      Mean |     Error |    StdDev |         Op/s | Scaled | ScaledSD | Allocated |
------------- |----------:|----------:|----------:|-------------:|-------:|---------:|----------:|
 TwoArguments | 330.32 ns | 4.2330 ns | 3.9595 ns |  3,027,379.5 |   5.94 |     0.10 |      88 B |
  NoArguments |  55.58 ns | 0.8043 ns | 0.7523 ns | 17,992,955.1 |   1.00 |     0.00 |       0 B |
```

This doesn't address scope impact but adds benchmarks to be able to iterate on perf further.