@pakrym 2018-10-26
I'm looking into TwoArguments_DefineMessage_FilteredByLevel slowdown, seems like a side-effect of having to allocate stack space for structs.