@pakrym 2018-10-26
Fixed regression, updated numbers.