@AndyAyersMS 2018-10-26
Trying to parse what you wrote there ... if you new up a struct conditionally and that struct contains GC types you may pay costs in the method prolog to have the GC parts of the struct zeroed even if you never end up using the struct.

That is if S is a struct with GC type fields, then

```C#
if (oftenNotTrue())
{
     use(new S(...));
}
```
will pay some per-call cost for the potential `new S` even if `oftenNotTrue()` is indeed mostly false. As you note splitting off the `new S` so it happens in a child frame will avoid this.

This is one of those "currently" true things because the jit isn't very smart about structs passed to calls (even if there's a behind the scenes copy)....

```asm
G_M63142_IG01:
       sub      rsp, 56
       vzeroupper 
       xor      rax, rax                    // prolog cost (always paid)
       mov      qword ptr [rsp+28H], rax    //  ... cont
       mov      qword ptr [rsp+30H], rax    // nulling 2 gc fields

G_M63142_IG02:
       call     F:OftenNotTrue():bool
       test     eax, eax
       je       SHORT G_M63142_IG03
       lea      rdx, bword ptr [rsp+28H]    // "new" S cost (conditionally paid)
       vxorps   xmm0, xmm0                  // ... cont
       vmovdqu  qword ptr [rdx], xmm0       // .. .cont
       mov      rcx, 0xD1FFAB1E
       call     F:Use(byref)

G_M63142_IG03:
       nop      

G_M63142_IG04:
       add      rsp, 56
       ret      
```